# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "所持ゲーム管理システムのバックエンドAPI仕様書",
        "title": "Yutagame Backend API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/app/game-recommendations": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public Recommendations"
                ],
                "summary": "ゲーム推薦を送信",
                "parameters": [
                    {
                        "description": "推薦内容",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.GameRecommendationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.GameRecommendation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "app.GameRecommendationRequest": {
            "type": "object",
            "properties": {
                "gameName": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.GameRecommendation": {
            "type": "object",
            "properties": {
                "adminNote": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "gameName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"Bearer {token}\" の形式でJWTトークンを入力してください。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}